Move option declations to picard/options.py and introduce const.defaults
