+++ picard/coverart/providers/local.py
@@ -25,10 +25,8 @@
 import os
 import re
 
-from picard.config import (
-    TextOption,
-    get_config,
-)
+from picard.config import get_config
+from picard.const.defaults import DEFAULT_LOCAL_COVER_ART_REGEX
 from picard.coverart.image import LocalFileCoverArtImage
 from picard.coverart.providers.provider import (
     CoverArtProvider,
@@ -46,10 +44,8 @@ class ProviderOptionsLocal(ProviderOptions):
     """
 
     HELP_URL = '/config/options_local_files.html'
-    _DEFAULT_LOCAL_COVER_ART_REGEX = r'^(?:cover|folder|albumart)(.*)\.(?:jpe?g|png|gif|tiff?|webp)$'
 
     options = [
-        TextOption('setting', 'local_cover_regex', _DEFAULT_LOCAL_COVER_ART_REGEX),
     ]
 
     _options_ui = Ui_LocalOptions
@@ -60,7 +56,7 @@ class ProviderOptionsLocal(ProviderOptions):
         self.ui.local_cover_regex_default.clicked.connect(self.set_local_cover_regex_default)
 
     def set_local_cover_regex_default(self):
-        self.ui.local_cover_regex_edit.setText(self._DEFAULT_LOCAL_COVER_ART_REGEX)
+        self.ui.local_cover_regex_edit.setText(DEFAULT_LOCAL_COVER_ART_REGEX)
 
     def load(self):
         config = get_config()
