+++ picard/ui/cdlookup.py
@@ -33,10 +33,7 @@ from PyQt6 import (
 )
 
 from picard import log
-from picard.config import (
-    Option,
-    get_config,
-)
+from picard.config import get_config
 from picard.i18n import gettext as _
 from picard.mbjson import (
     artist_credit_from_node,
@@ -57,7 +54,6 @@ class CDLookupDialog(PicardDialog):
     dialog_header_state = 'cdlookupdialog_header_state'
 
     options = [
-        Option('persist', dialog_header_state, QtCore.QByteArray())
     ]
 
     def __init__(self, releases, disc, parent=None):
