+++ picard/ui/mainwindow.py
@@ -69,12 +69,7 @@ from picard.cluster import (
     FileList,
 )
 from picard.config import (
-    BoolOption,
-    FloatOption,
-    IntOption,
-    Option,
     SettingConfigSection,
-    TextOption,
     get_config,
 )
 from picard.const import PROGRAM_UPDATE_LEVELS
@@ -182,15 +177,6 @@ class MainWindow(QtWidgets.QMainWindow, PreserveGeometry):
     ready_for_display = QtCore.pyqtSignal()
 
     options = [
-        Option('persist', 'window_state', QtCore.QByteArray()),
-        BoolOption('persist', 'window_maximized', False),
-        BoolOption('persist', 'view_metadata_view', True),
-        BoolOption('persist', 'view_cover_art', True),
-        BoolOption('persist', 'view_toolbar', True),
-        BoolOption('persist', 'view_file_browser', False),
-        TextOption('persist', 'current_directory', ""),
-        FloatOption('persist', 'mediaplayer_playback_rate', 1.0),
-        IntOption('persist', 'mediaplayer_volume', 50),
     ]
 
     def __init__(self, parent=None, disable_player=False):
