+++ picard/ui/options/cover.py
@@ -27,13 +27,9 @@
 
 
 from picard.config import (
-    BoolOption,
-    ListOption,
     Option,
-    TextOption,
     get_config,
 )
-from picard.const import DEFAULT_COVER_IMAGE_FILENAME
 from picard.coverart.providers import cover_art_providers
 from picard.i18n import (
     N_,
@@ -50,14 +46,6 @@ from picard.ui.ui_options_cover import Ui_CoverOptionsPage
 from picard.ui.util import qlistwidget_items
 
 
-DEFAULT_CA_PROVIDERS = [
-    ('Cover Art Archive', True),
-    ('UrlRelationships', True),
-    ('CaaReleaseGroup', True),
-    ('Local', False),
-]
-
-
 class CoverOptionsPage(OptionsPage):
 
     NAME = 'cover'
@@ -68,14 +56,6 @@ class CoverOptionsPage(OptionsPage):
     HELP_URL = "/config/options_cover.html"
 
     options = [
-        BoolOption('setting', 'save_images_to_tags', True, title=N_("Embed cover images into tags")),
-        BoolOption('setting', 'embed_only_one_front_image', True, title=N_("Embed only a single front image")),
-        BoolOption('setting', 'save_images_to_files', False, title=N_("Save cover images as separate files")),
-        TextOption('setting', 'cover_image_filename', DEFAULT_COVER_IMAGE_FILENAME, title=N_("File name for images")),
-        BoolOption('setting', 'save_images_overwrite', False, title=N_("Overwrite existing image files")),
-        BoolOption('setting', 'save_only_one_front_image', False, title=N_("Save only a single front image as separate file")),
-        BoolOption('setting', 'image_type_as_filename', False, title=N_("Always use the primary image type as the file name for non-front images")),
-        ListOption('setting', 'ca_providers', DEFAULT_CA_PROVIDERS, title=N_("Cover art providers")),
     ]
 
     def __init__(self, parent=None):
