+++ picard/ui/options/interface_colors.py
@@ -28,17 +28,13 @@ from PyQt6 import (
     QtWidgets,
 )
 
-from picard.config import Option
 from picard.const.sys import IS_MACOS
 from picard.i18n import (
     N_,
     gettext as _,
 )
 
-from picard.ui.colors import (
-    InterfaceColors,
-    interface_colors,
-)
+from picard.ui.colors import interface_colors
 from picard.ui.options import (
     OptionsPage,
     register_options_page,
@@ -101,8 +97,6 @@ class InterfaceColorsOptionsPage(OptionsPage):
     HELP_URL = "/config/options_interface_colors.html"
 
     options = [
-        Option('setting', 'interface_colors', InterfaceColors(dark_theme=False).get_colors(), title=N_("Colors to use for light theme")),
-        Option('setting', 'interface_colors_dark', InterfaceColors(dark_theme=True).get_colors(), title=N_("Colors to use for dark theme")),
     ]
 
     def __init__(self, parent=None):
