+++ picard/ui/options/tags_compatibility_aac.py
@@ -21,10 +21,7 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 
-from picard.config import (
-    BoolOption,
-    get_config,
-)
+from picard.config import get_config
 from picard.i18n import N_
 
 from picard.ui.options import (
@@ -46,8 +43,6 @@ class TagsCompatibilityAACOptionsPage(OptionsPage):
     HELP_URL = "/config/options_tags_compatibility_aac.html"
 
     options = [
-        BoolOption('setting', 'aac_save_ape', True, title=N_("Save APEv2 tags to AAC")),
-        BoolOption('setting', 'remove_ape_from_aac', False, title=N_("Remove APEv2 tags from AAC files")),
     ]
 
     def __init__(self, parent=None):
