+++ picard/ui/searchdialog/artist.py
@@ -21,12 +21,7 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 
-from PyQt6 import QtCore
-
-from picard.config import (
-    Option,
-    get_config,
-)
+from picard.config import get_config
 from picard.i18n import gettext as _
 from picard.mbjson import artist_to_metadata
 from picard.metadata import Metadata
@@ -42,7 +37,6 @@ class ArtistSearchDialog(SearchDialog):
     dialog_header_state = 'artistsearchdialog_header_state'
 
     options = [
-        Option('persist', dialog_header_state, QtCore.QByteArray())
     ]
 
     def __init__(self, parent):
