+++ picard/ui/tagsfromfilenames.py
@@ -32,10 +32,7 @@ import re
 
 from PyQt6 import QtWidgets
 
-from picard.config import (
-    TextOption,
-    get_config,
-)
+from picard.config import get_config
 from picard.i18n import gettext as _
 from picard.script.parser import normalize_tagname
 from picard.util.tags import display_tag_name
@@ -102,7 +99,6 @@ class TagsFromFileNamesDialog(PicardDialog):
     help_url = 'doc_tags_from_filenames'
 
     options = [
-        TextOption('persist', 'tags_from_filenames_format', ''),
     ]
 
     def __init__(self, files, parent=None):
